PP version of braitenberg vehicle.
Inference should be implemented and working, found a mistake in previous account (not recorded).

Noise and random fluctuations are included and the systems seems to infer decently.

Time to implement some active inference, with ideally a prior on a level of light to be attained (imagine a natural behaviour towards the light stopped by action, or vice versa)